catalog error handining

+++ store-pod/catalog/catalog-core/src/main/java/com/asrevo/cvhome/catalog/service/facade/product/ProductOptionSetFacadeImpl.java
@@ -5,6 +5,9 @@ import java.util.List;
 import org.springframework.stereotype.Service;
 
 import com.asrevo.cvhome.catalog.entity.product.attribute.ProductOptionSet;
+import com.asrevo.cvhome.catalog.errors.DuplicateProductOptionSetException;
+import com.asrevo.cvhome.catalog.errors.ProductOptionSetNotFoundException;
+import com.asrevo.cvhome.catalog.errors.ProductTypeNotFoundException;
 import com.asrevo.cvhome.catalog.model.product.attribute.optionset.PersistableProductOptionSet;
 import com.asrevo.cvhome.catalog.model.product.attribute.optionset.ReadableProductOptionSet;
 import com.asrevo.cvhome.catalog.model.product.type.ReadableProductType;
@@ -13,17 +16,10 @@ import com.asrevo.cvhome.catalog.service.mapper.catalog.ReadableProductOptionSet
 import com.asrevo.cvhome.catalog.services.product.attribute.ProductOptionSetService;
 import com.asrevo.cvhome.commons.domain.LanguageCode;
 import com.asrevo.cvhome.commons.domain.StoreMerchantId;
-import com.asrevo.cvhome.store.controller.exception.OperationNotAllowedException;
-import com.asrevo.cvhome.store.controller.exception.ResourceNotFoundException;
-import com.asrevo.cvhome.store.controller.exception.ServiceRuntimeException;
 import com.asrevo.cvhome.store.core.exception.ServiceException;
 
 @Service
 public class ProductOptionSetFacadeImpl implements ProductOptionSetFacade {
-
-    private static final String PRODUCT_OPTION_SET_NOT_FOUND_TEMPLATE = "ProductOptionSet not found for id [%s] and store [%s]";
-
-    private static final String CANT_FIND_PRODUCT_TYPE_TEMPLATE = "Can't fing product type [%s] fpr merchand [%s]";
 
     private final PersistableProductOptionSetMapper persistableProductOptionSetMapper;
 
@@ -43,10 +39,11 @@ public class ProductOptionSetFacadeImpl implements ProductOptionSetFacade {
     }
 
     @Override
-    public ReadableProductOptionSet get(Long id, StoreMerchantId store, LanguageCode language) {
+    public ReadableProductOptionSet get(Long id, StoreMerchantId store, LanguageCode language)
+            throws ProductOptionSetNotFoundException {
         ProductOptionSet optionSet = productOptionSetService.getById(store, id, language);
         if (optionSet == null) {
-            throw new ResourceNotFoundException(PRODUCT_OPTION_SET_NOT_FOUND_TEMPLATE.formatted(id, store));
+            throw ProductOptionSetNotFoundException.of(id, store);
         }
 
         return readableProductOptionSetMapper.convert(optionSet, store, language);
@@ -63,26 +60,23 @@ public class ProductOptionSetFacadeImpl implements ProductOptionSetFacade {
     }
 
     @Override
-    public void create(PersistableProductOptionSet optionSet, StoreMerchantId store, LanguageCode language) {
+    public void create(PersistableProductOptionSet optionSet, StoreMerchantId store, LanguageCode language)
+            throws DuplicateProductOptionSetException, ServiceException {
         if (this.exists(optionSet.getCode(), store)) {
-            throw new OperationNotAllowedException(
-                    "Option set with code [%s] already exist".formatted(optionSet.getCode()));
+            throw DuplicateProductOptionSetException.of(optionSet.getCode(), store);
         }
 
         ProductOptionSet opt = persistableProductOptionSetMapper.convert(optionSet, store, language);
-        try {
-            opt.setStoreMerchantId(store);
-            productOptionSetService.create(opt);
-        } catch (ServiceException e) {
-            throw new ServiceRuntimeException("Exception while creating ProductOptionSet", e);
-        }
+        opt.setStoreMerchantId(store);
+        productOptionSetService.create(opt);
     }
 
     @Override
-    public void update(Long id, PersistableProductOptionSet optionSet, StoreMerchantId store, LanguageCode language) {
+    public void update(Long id, PersistableProductOptionSet optionSet, StoreMerchantId store, LanguageCode language)
+            throws ProductOptionSetNotFoundException {
         ProductOptionSet opt = productOptionSetService.getById(store, id, language);
         if (opt == null) {
-            throw new ResourceNotFoundException(PRODUCT_OPTION_SET_NOT_FOUND_TEMPLATE.formatted(id, store));
+            throw ProductOptionSetNotFoundException.of(id, store);
         }
 
         optionSet.setId(id);
@@ -93,19 +87,12 @@ public class ProductOptionSetFacadeImpl implements ProductOptionSetFacade {
     }
 
     @Override
-    public void delete(Long id, StoreMerchantId store) {
+    public void delete(Long id, StoreMerchantId store) throws ProductOptionSetNotFoundException, ServiceException {
         ProductOptionSet opt = productOptionSetService.getById(id);
-        if (opt == null) {
-            throw new ResourceNotFoundException(PRODUCT_OPTION_SET_NOT_FOUND_TEMPLATE.formatted(id, store));
+        if (opt == null || !opt.getStoreMerchantId().equals(store)) {
+            throw ProductOptionSetNotFoundException.of(id, store);
         }
-        if (!opt.getStoreMerchantId().equals(store)) {
-            throw new ResourceNotFoundException(PRODUCT_OPTION_SET_NOT_FOUND_TEMPLATE.formatted(id, store));
-        }
-        try {
-            productOptionSetService.delete(opt);
-        } catch (ServiceException e) {
-            throw new ServiceRuntimeException("Exception while deleting ProductOptionSet", e);
-        }
+        productOptionSetService.delete(opt);
     }
 
     @Override
@@ -115,12 +102,10 @@ public class ProductOptionSetFacadeImpl implements ProductOptionSetFacade {
     }
 
     @Override
-    public List<ReadableProductOptionSet> list(StoreMerchantId store, LanguageCode language, String type) {
+    public List<ReadableProductOptionSet> list(StoreMerchantId store, LanguageCode language, String type)
+            throws ProductTypeNotFoundException {
+        // productTypeFacade.get now throws rather than returning null, so the null check that followed it is gone.
         ReadableProductType readable = productTypeFacade.get(store, type, language);
-
-        if (readable == null) {
-            throw new ResourceNotFoundException(CANT_FIND_PRODUCT_TYPE_TEMPLATE.formatted(type, store));
-        }
 
         List<ProductOptionSet> optionSets = productOptionSetService.getByProductType(readable.getId(), store, language);
         return optionSets.stream().map(opt -> this.convert(opt, store, language)).toList();
